apply version number for backward compatibility & create YFGUtils to avoid duplicate functions or methods
